기능추가: 파일 분리 및 네이밍 수정
